+++ backend/src/main/kotlin/org/aleks616/shrendar/band/model/ArtistBandAddDto.kt
@@ -3,9 +3,9 @@ package org.aleks616.shrendar.band.model
 import java.io.Serializable
 
 data class ArtistBandAddDto(
-    var id:Int?=null,
+    var id:Long?=null,
     var bandId:Int?=null,
-    var artistId:Int?=null,
+    var artistId:Long?=null,
     var nickname:String?=null,
     var role:String?=null,
     var joinedYear:Int?=null,
