+++ backend/src/main/kotlin/org/aleks616/shrendar/band/model/ArtistBandsExtendedDto.kt
@@ -5,7 +5,7 @@ import java.io.Serializable
 
 data class ArtistBandsExtendedDto(
     val id:Int?=null,
-    val artistId:Int?=null,
+    val artistId:Long?=null,
     val artistName:String?=null,
     val bandId:Int?=null,
     val bandName:String?=null,
