+++ backend/src/main/kotlin/org/aleks616/shrendar/band/model/BandsMembers.kt
@@ -17,7 +17,7 @@ open class BandsMembers {
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
     @Column(name="id",nullable=false)
-    open var id: Int? = null
+    open var id:Long?=null
 
     @ManyToOne(fetch=FetchType.EAGER)
     @JoinColumn(name="artist_id")
