+++ backend/src/main/kotlin/org/aleks616/shrendar/band/model/BandsMembersDataDto.kt
@@ -6,8 +6,8 @@ import java.io.Serializable
  * DTO for {@link org.aleks616.shrendar.band.model.BandsMembers}
  */
 data class BandsMembersDataDto(
-    val id:Int?=null,
-    val artistId:Int?=null,
+    val id:Long?=null,
+    val artistId:Long?=null,
     val artistName:String?=null,
     val bandId:Int?=null,
     val bandName:String?=null,
